use completely clean edge function file
